refactor: 래거시 MVC 제거
