+++ README.md
@@ -35,7 +35,7 @@
 
 - [x] JspView 클래스 구현
 - [x] JsonView 클래스 구현
-- [ ] Legacy MVC 제거
+- [x] Legacy MVC 제거
 
 리팩터링
