+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -17,8 +17,6 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
         log.info("Start AppWebApplication Initializer");
 
         final var dispatcherServlet = new DispatcherServlet();
-        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
-        dispatcherServlet.addHandlerAdapter(new ManualHandlerAdapter());
         dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse"));
         dispatcherServlet.addHandlerAdapter(new AnnotationHandlerAdapter());
